Move PaymentMethodStep to own file and convert to TS
